requirements.txt
streamlit>=1.35.0
anthropic>=0.30.0
google-genai>=1.0.0
groq>=0.9.0
python-docx>=1.1.0
python-dotenv>=1.0.0
